Feat: defaultApi.js 수정
